crud products

+++ frontend-angular/src/app/features/admin/categories/category-list/category-list.component.ts
@@ -1,16 +1,17 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
 import { RouterLink } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
 
 import { ICategory } from '../../../../core/models/categories.model';
 import { PUBLIC_ENDPOINTS } from '../../../../core/constants/constants';
 import { CategoriesService } from '../../../../core/services/categories.service';
-import { ToastrService } from 'ngx-toastr';
+import { LoaderComponent } from '../../../../shared/ui/loader/loader.component';
 
 @Component({
   selector: 'app-admin-category-list',
   standalone: true,
-  imports: [MatIconModule, RouterLink],
+  imports: [MatIconModule, RouterLink, LoaderComponent],
   templateUrl: './category-list.component.html',
   styleUrl: './category-list.component.css'
 })
@@ -18,10 +19,11 @@ import { ToastrService } from 'ngx-toastr';
 export class CategoryListComponent implements OnInit {
   categories: ICategory[] = [];
   imageUrl = PUBLIC_ENDPOINTS.IMAGE_CATEGORIES;
+  isLoading: boolean = false;
 
   constructor(
     private categoriesService: CategoriesService,
-    private toast: ToastrService
+    private toast: ToastrService,
   ) { }
 
   ngOnInit(): void {
@@ -29,18 +31,25 @@ export class CategoryListComponent implements OnInit {
   }
 
   getAllCategories() {
+    this.isLoading = true;
     this.categoriesService.getAllCategories().subscribe({
       next: (response) => {
         this.categories = response.data;
+        setTimeout(() => {
+          this.isLoading = false;
+        }, 500);
       }
     });
   }
 
   onDeleteCategory(id: string) {
-    this.categoriesService.deleteCategory(id).subscribe({
-      next: (response) => {
-        this.toast.success('Category deleted successfully!');
-      }
-    });
+    if (confirm('Do you want to delete this category?')) {
+      this.categoriesService.deleteCategory(id).subscribe({
+        next: (response) => {
+          this.toast.success('Category deleted successfully!');
+          this.getAllCategories();
+        }
+      });
+    }
   }
 }
